# [tool] Update baseline data and optimize test time for RegressionChecker tool.