| Platform | Before optimizing | After optimizing |
| --- | --- | --- |
| `MacOS`| ~230s | ~123s |
| `Linux` | ~151s | ~91s |
| `Windows` | ~215s | ~138s |
| `Android` | ~425s | ~224s |

@ibelem PTAL, thanks.